style: fix clippy warnings
